Move Authorization to it's own module
