+++ src/middleware/authorization/same_user_auth.rs
@@ -1,41 +1,10 @@
 use iron::prelude::*;
-use iron::BeforeMiddleware;
 use iron::status::Status;
 use iron_login::User as U;
 
 use error::UnauthorizedError;
 use models::user::{self, User};
-
-#[derive(Clone, Debug)]
-pub struct Authorizer<T: Send + Sync> {
-    reqs: Vec<T>
-}
-
-impl<T: Send + Sync> Authorizer<T> {
-    pub fn new(r: Vec<T>) -> Authorizer<T> {
-        Authorizer {
-            reqs: r
-        }
-    }
-}
-
-impl<T: UserRequirement + Send + Sync + 'static> BeforeMiddleware for Authorizer<T> {
-    fn before(&self, req: &mut Request) -> IronResult<()> {
-        let user = User::get_login(req).get_user();
-
-        let results = self.reqs.iter().map(|x| x.check(user.as_ref(), req)).all(|x| x);
-
-        if results {
-            Ok(())
-        } else {
-            Err(IronError::new(UnauthorizedError::new(user), Status::Unauthorized))
-        }
-    }
-}
-
-pub trait UserRequirement {
-    fn check(&self, user: Option<&User>, req: &mut Request) -> bool;
-}
+use middleware::authorization::UserRequirement;
 
 #[derive(Clone, Debug)]
 pub struct SameUserAuth;
@@ -67,3 +36,4 @@ impl UserRequirement for SameUserAuth {
         other_user.map(|x| x.id == user.id).unwrap_or(false)
     }
 }
+
